Add files via upload
Automatically creates a log as soon as it is imported with a time stamp in the log file name if not already open. Although the Python libraries automatically close all open file handles upon exit, there is a close_log() method to flush and close the file. This is not only useful for tradditional programmers who want a greater degree of program control, but useful in conjunction with control programs such as Ansible whereby printing to STD_OUT needs to be suppressed for all log messages except the final completion message.